requirements.txt
fastapi
uvicorn
requests
numpy
torch
transformers
python-dotenv
scikit-learn
